Made Exception fit to be shown in output.

+++ src/SiteContent/ApiCall/DeletePageApiCall.php
@@ -5,6 +5,7 @@ namespace Spot\SiteContent\ApiCall;
 use Psr\Http\Message\RequestInterface as HttpRequest;
 use Psr\Http\Message\ServerRequestInterface as ServerHttpRequest;
 use Psr\Log\LoggerInterface;
+use Psr\Log\LogLevel;
 use Ramsey\Uuid\Uuid;
 use Spot\Api\LoggableTrait;
 use Spot\Api\Request\Executor\ExecutorInterface;
@@ -13,6 +14,8 @@ use Spot\Api\Request\Message\ArrayRequest;
 use Spot\Api\Request\Message\RequestInterface;
 use Spot\Api\Response\Message\ArrayResponse;
 use Spot\Api\Response\Message\ResponseInterface;
+use Spot\Api\Response\Message\ServerErrorResponse;
+use Spot\Api\Response\ResponseException;
 use Spot\Common\ParticleFixes\Validator;
 use Spot\Common\Request\ValidationFailedException;
 use Spot\SiteContent\Repository\PageRepository;
@@ -47,8 +50,18 @@ class DeletePageApiCall implements HttpRequestParserInterface, ExecutorInterface
 
     public function executeRequest(RequestInterface $request, HttpRequest $httpRequest) : ResponseInterface
     {
-        $page = $this->pageRepository->getByUuid(Uuid::fromString($request['uuid']));
-        $this->pageRepository->delete($page);
-        return new ArrayResponse(self::MESSAGE, ['data' => $page]);
+        if (!$request instanceof ArrayRequest) {
+            $this->log(LogLevel::ERROR, 'Did not receive an ArrayRequest instance.');
+            throw new ResponseException('An error occurred during DeletePageApiCall.', new ServerErrorResponse());
+        }
+
+        try {
+            $page = $this->pageRepository->getByUuid(Uuid::fromString($request['uuid']));
+            $this->pageRepository->delete($page);
+            return new ArrayResponse(self::MESSAGE, ['data' => $page]);
+        } catch (\Throwable $e) {
+            $this->log(LogLevel::ERROR, $e->getMessage());
+            throw new ResponseException('An error occurred during DeletePageApiCall.', new ServerErrorResponse());
+        }
     }
 }
